Main Carbon class created
Created the main carbon class which will handle the GUI and rendering of the world.
I also changed the Vec.h Vector-Vector assignment operataors to work like the Vector-scalar operators.
The main.cpp file was also edited to include basic World and Carbon class and noted that the default GUI code should be removed.