Fixed Vec.h
I did another oopsie(Sorry I am new to c++). In the assignment operator overloading I accidently put semi-colon on the outside of the brackets. For some reason this was successfully building and passing on Windows but not on linux or macos.
This should hopefully make it so all builds pass now(hopefully).