Updated README & Some collsion stuff
I have updated the README with features that I plan to include in this project.
On another note, I went back to this article https://winter.dev/articles/physics-engine
and found that the collsion detection seemed not great as it did not use things like AABB
so I plan to stray away from that and go to my own method.